chore: split db/mod.rs into domain sub-modules
2026-line monolith split into journal.rs, analytics.rs, books.rs,
signals.rs, and voice_memos.rs. mod.rs retains only Database::new()
(all migrations), get_db_path, and auth helpers. All public types
re-exported via pub use so call sites are unchanged.
